+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Add animations toggle
+- Add some animations
 
 ## v0.9.0 - 2026-02-20